// frontend/lib/api/models/PlaylistUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PlaylistUpdate = {
    title?: (string | null);
    description?: (string | null);
    thumbnail?: (string | null);
    is_public?: (boolean | null);
    category?: (string | null);
    scenario?: (string | null);
    vibes?: (Array<string> | null);
    hook_description?: (string | null);
};